Разбила код на модули - подключила их в главном модуле - подключила главный модуль на страницу index
